refactor excel dn tests
